refactor(test): P5-2 follow-up — extract CSS filter chain cluster
Continue test_canvas.cpp split. Extracts the 377-line pulp #1434
Phase A2-4 cluster into a sibling TU. Two sub-clusters shipped
together since both pin the same backend contract:

  1. Skia-only pixel readback (-219 lines): contrast(0) renders
     ~mid-gray; invert(1) maps black → white; opacity ordering
     with drop-shadow changes pixel output; set_filter parses
     drop-shadow with hex color + parses fallback null.

  2. Portable filter-chain matrix math (-158 lines): contrast(0)
     bias matrix lands at mid-gray (~128); invert(1) maps
     black→white via the matrix; invert(0) is identity; opacity(α)
     scales alpha and preserves RGB.

Extracted file preserves the #ifdef PULP_HAS_SKIA / #endif pairing
around the pixel-readback subset; parent test_canvas.cpp's
#ifdef structure (lines 667-1043) remains intact.

* test_canvas.cpp: 2,735 → 2,358 lines (-377)
* New target pulp-test-canvas-filter-chain: 421 lines, 10
  TEST_CASEs, 52 assertions (8 pass + 2 expected-fail).
